Fix use integer division inside floating-point computation

+++ org.eventb.core/src/org/eventb/internal/core/pom/AutoPOM.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2005, 2013 ETH Zurich and others.
+ * Copyright (c) 2005, 2017 ETH Zurich and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -120,7 +120,7 @@ public class AutoPOM implements IAutomaticTool, IExtractor {
 				}
 				else{
 				proofReuse = ((PSUpdater.recoverablePOsWithProofs + PSUpdater.unchangedPOsWithProofs) * 100)
-								 / (PSUpdater.totalPOs - PSUpdater.newPOs);
+								 / (float) (PSUpdater.totalPOs - PSUpdater.newPOs);
 				}
 				
 				System.out.println(
